# OGCServiceCatalogue/Models/GeneralResponse.cs
/*
 * Edge Connection Manager
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace OGCServiceCatalogue.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class GeneralResponse : IEquatable<GeneralResponse>
    { 
        /// <summary>
        /// returns 1 if successful
        /// </summary>
        /// <value>returns 1 if successful</value>
        [Required]
        [DataMember(Name="success")]
        public bool? Success { get; set; }

        /// <summary>
        /// a short comment
        /// </summary>
        /// <value>a short comment</value>
        [Required]
        [DataMember(Name="description")]
        public string Description { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class GeneralResponse {\n");
            sb.Append("  Success: ").Append(Success).Append("\n");
            sb.Append("  Description: ").Append(Description).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((GeneralResponse)obj);
        }

        /// <summary>
        /// Returns true if GeneralResponse instances are equal
        /// </summary>
        /// <param name="other">Instance of GeneralResponse to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(GeneralResponse other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Success == other.Success ||
                    Success != null &&
                    Success.Equals(other.Success)
                ) && 
                (
                    Description == other.Description ||
                    Description != null &&
                    Description.Equals(other.Description)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Success != null)
                    hashCode = hashCode * 59 + Success.GetHashCode();
                    if (Description != null)
                    hashCode = hashCode * 59 + Description.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(GeneralResponse left, GeneralResponse right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(GeneralResponse left, GeneralResponse right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}